change reindexing approach
