+++ api/services/DocumentationIndex.js
@@ -1,37 +1,14 @@
 const fs = require('fs').promises;
 const path = require('path');
-const { marked } = require('marked');
-const cheerio = require('cheerio');
 const Fuse = require('fuse.js');
-const natural = require('natural');
 const logger = require('../utils/logger');
 
-// Only import chokidar in non-serverless environments
-let chokidar;
-const isServerless = process.env.NETLIFY || process.env.AWS_LAMBDA_FUNCTION_NAME || process.env.VERCEL;
-if (!isServerless) {
-    try {
-        chokidar = require('chokidar');
-    } catch (error) {
-        logger.warn('chokidar not available - file watching disabled');
-    }
-}
 
 class DocumentationIndex {
     constructor(options = {}) {
-        this.contentPath = options.contentPath || this.resolveContentPath();
         this.indexData = [];
         this.fuseIndex = null;
         this.topics = new Map();
-        this.watcher = null;
-        this.rebuildTimeout = null;
-        
-        // Configure marked for parsing markdown
-        marked.setOptions({
-            gfm: true,
-            breaks: false,
-            smartLists: true
-        });
         
         // Configure Fuse.js options for fuzzy search
         this.fuseOptions = {
@@ -49,257 +26,48 @@ class DocumentationIndex {
         };
     }
 
-    resolveContentPath() {
-        // Log environment info for debugging
-        logger.info(`🔍 RESOLVING CONTENT PATH:`);
-        logger.info(`   - Node ENV: ${process.env.NODE_ENV}`);
-        logger.info(`   - NETLIFY: ${process.env.NETLIFY}`);
-        logger.info(`   - Working dir: ${process.cwd()}`);
-        logger.info(`   - __dirname: ${__dirname}`);
-        
-        // Two main scenarios to handle:
-        const possiblePaths = [
-            // 1. Local development: running from api/services/ directory
-            path.join(__dirname, '../../content/en'),
-            // 2. Serverless/production: working directory is project root
-            path.join(process.cwd(), 'content/en')
-        ];
 
-        logger.info(`   - Testing paths: ${possiblePaths.join(', ')}`);
-
-        // Try each path and use the first one that exists
-        for (const contentPath of possiblePaths) {
-            try {
-                const fs = require('fs');
-                logger.info(`   - Testing: ${contentPath}`);
-                if (fs.existsSync(contentPath)) {
-                    logger.info(`✅ Content path resolved to: ${contentPath}`);
-                    return contentPath;
+    async loadPrebuiltIndex() {
+        try {
+            const indexPath = path.join(process.cwd(), 'search-index.json');
+            const indexData = await fs.readFile(indexPath, 'utf-8');
+            const parsedData = JSON.parse(indexData);
+            
+            // Load the documents into our index
+            this.indexData = parsedData.documents;
+            
+            // Rebuild topics map
+            this.topics.clear();
+            for (const doc of this.indexData) {
+                if (doc.topic) {
+                    if (!this.topics.has(doc.topic)) {
+                        this.topics.set(doc.topic, []);
+                    }
+                    this.topics.get(doc.topic).push(doc);
                 }
-                logger.warn(`❌ Path not found: ${contentPath}`);
-            } catch (error) {
-                logger.error(`❌ Path ${contentPath} not accessible: ${error.message}`);
             }
+            
+            this.createFuseIndex();
+            return true;
+        } catch (error) {
+            logger.error(`Failed to load pre-built index: ${error.message}`);
+            return false;
         }
-
-        // If neither works, log diagnostic info and use serverless path as fallback
-        const fallbackPath = path.join(process.cwd(), 'content/en');
-        logger.error(`🚨 CONTENT PATH RESOLUTION FAILED!`);
-        logger.error(`Working directory: ${process.cwd()}`);
-        logger.error(`__dirname: ${__dirname}`);
-        logger.error(`Using fallback: ${fallbackPath}`);
-        return fallbackPath;
     }
 
     async initialize() {
         try {
-            logger.info('🚀 Initializing documentation index...');
-            logger.info(`   Environment: ${process.env.NODE_ENV || 'unknown'}`);
-            logger.info(`   Serverless: ${process.env.NETLIFY ? 'YES (Netlify)' : 'NO'}`);
-            
-            await this.buildIndex();
-            this.createFuseIndex();
-            
-            logger.info(`✅ Documentation index built with ${this.indexData.length} documents`);
-            logger.info(`📂 Topics discovered: ${Array.from(this.topics.keys()).join(', ')}`);
+            const loaded = await this.loadPrebuiltIndex();
+            if (!loaded) {
+                throw new Error('Pre-built search index not found');
+            }
+            logger.info(`Loaded ${this.indexData.length} documents`);
         } catch (error) {
-            logger.error('❌ Failed to initialize documentation index:', error);
-            logger.error('Stack trace:', error.stack);
+            logger.error('Failed to initialize documentation index:', error);
             throw error;
         }
     }
 
-    async buildIndex() {
-        this.indexData = [];
-        this.topics.clear();
-        
-        logger.info(`Starting to build index from content path: ${this.contentPath}`);
-        
-        // Check if content path exists
-        try {
-            await fs.access(this.contentPath);
-            logger.info(`Content path is accessible: ${this.contentPath}`);
-            
-            // Debug: List directory contents
-            const entries = await fs.readdir(this.contentPath, { withFileTypes: true });
-            logger.info(`Directory contents (${entries.length} entries):`, entries.map(e => `${e.name}${e.isDirectory() ? '/' : ''}`).join(', '));
-        } catch (error) {
-            logger.error(`Content path is not accessible: ${this.contentPath}`, error);
-            return; // Exit early if path doesn't exist
-        }
-        
-        await this.processDirectory(this.contentPath);
-        
-        // Debug: Log final results
-        logger.info(`Build complete: ${this.indexData.length} documents processed`);
-    }
-
-    async processDirectory(dirPath) {
-        try {
-            logger.debug(`📁 Processing directory: ${dirPath}`);
-            const entries = await fs.readdir(dirPath, { withFileTypes: true });
-            logger.debug(`   Found ${entries.length} entries`);
-            
-            let markdownCount = 0;
-            let dirCount = 0;
-            
-            for (const entry of entries) {
-                const fullPath = path.join(dirPath, entry.name);
-                
-                if (entry.isDirectory()) {
-                    dirCount++;
-                    logger.debug(`   📁 Subdirectory: ${entry.name}`);
-                    await this.processDirectory(fullPath);
-                } else if (entry.isFile() && entry.name.endsWith('.md')) {
-                    markdownCount++;
-                    logger.debug(`   📄 Markdown file: ${entry.name}`);
-                    await this.processMarkdownFile(fullPath);
-                }
-            }
-            
-            logger.debug(`📊 Directory ${dirPath}: ${markdownCount} markdown files, ${dirCount} subdirectories`);
-        } catch (error) {
-            logger.error(`❌ Error processing directory ${dirPath}:`, error);
-        }
-    }
-
-    async processMarkdownFile(filePath) {
-        try {
-            logger.debug(`Processing markdown file: ${filePath}`);
-            const content = await fs.readFile(filePath, 'utf-8');
-            const parsed = this.parseMarkdown(content);
-            
-            if (parsed && parsed.title) {
-                // Generate URL from file path
-                const url = this.generateUrl(filePath);
-                
-                // Extract topic from path
-                const topic = this.extractTopic(filePath);
-                
-                const document = {
-                    id: filePath,
-                    title: parsed.title,
-                    description: parsed.description || '',
-                    content: parsed.content,
-                    tags: parsed.tags || [],
-                    topic: topic,
-                    url: url,
-                    lastModified: (await fs.stat(filePath)).mtime,
-                    wordCount: this.countWords(parsed.content)
-                };
-                
-                this.indexData.push(document);
-                
-                // Add to topics map
-                if (topic) {
-                    if (!this.topics.has(topic)) {
-                        this.topics.set(topic, []);
-                    }
-                    this.topics.get(topic).push(document);
-                }
-                
-                logger.info(`✅ Indexed: ${document.title} (${document.wordCount} words) - ${document.url}`);
-            } else {
-                logger.warn(`⚠️  Failed to parse or no title found: ${filePath}`);
-            }
-        } catch (error) {
-            logger.error(`❌ Error processing file ${filePath}:`, error);
-        }
-    }
-
-    parseMarkdown(content) {
-        try {
-            // Extract frontmatter
-            const frontmatterMatch = content.match(/^---\n([\s\S]*?)\n---/);
-            let frontmatter = {};
-            let markdownContent = content;
-            
-            if (frontmatterMatch) {
-                markdownContent = content.substring(frontmatterMatch[0].length);
-                try {
-                    // Simple YAML parsing for common frontmatter fields
-                    const yamlLines = frontmatterMatch[1].split('\n');
-                    for (const line of yamlLines) {
-                        const colonIndex = line.indexOf(':');
-                        if (colonIndex > 0) {
-                            const key = line.substring(0, colonIndex).trim();
-                            const value = line.substring(colonIndex + 1).trim().replace(/^["']|["']$/g, '');
-                            frontmatter[key] = value;
-                        }
-                    }
-                } catch (yamlError) {
-                    logger.warn('Failed to parse frontmatter YAML:', yamlError);
-                }
-            }
-            
-            // Convert markdown to HTML
-            const html = marked(markdownContent);
-            
-            // Extract text content using cheerio
-            const $ = cheerio.load(html);
-            const textContent = $.text().replace(/\s+/g, ' ').trim();
-            
-            // Extract title (from frontmatter or first heading)
-            let title = frontmatter.title;
-            if (!title) {
-                const firstHeading = $('h1, h2').first().text().trim();
-                title = firstHeading || 'Untitled';
-            }
-            
-            // Extract description
-            let description = frontmatter.description;
-            if (!description) {
-                // Use first paragraph as description
-                const firstParagraph = $('p').first().text().trim();
-                description = firstParagraph.substring(0, 200) + (firstParagraph.length > 200 ? '...' : '');
-            }
-            
-            // Extract tags
-            let tags = [];
-            if (frontmatter.tags) {
-                tags = frontmatter.tags.split(',').map(tag => tag.trim());
-            }
-            
-            return {
-                title,
-                description,
-                content: textContent,
-                tags,
-                linkTitle: frontmatter.linkTitle || title,
-                weight: parseInt(frontmatter.weight) || 0
-            };
-            
-        } catch (error) {
-            logger.error('Error parsing markdown:', error);
-            return null;
-        }
-    }
-
-    generateUrl(filePath) {
-        // Convert file path to Hugo URL structure
-        // contentPath is now 'content/en', so get relative path from there
-        const relativePath = path.relative(this.contentPath, filePath);
-        let url = '/' + relativePath.replace(/\\/g, '/').replace(/\.md$/, '/').replace(/_index\/$/, '');
-        
-        url = url.replace(/\/+/g, '/');
-        
-        return url || '/';
-    }
-
-    extractTopic(filePath) {
-        const relativePath = path.relative(this.contentPath, filePath);
-        const pathParts = relativePath.split(path.sep);
-        
-        if (pathParts.length > 0) {
-            return pathParts[0];
-        }
-        
-        return 'general';
-    }
-
-    countWords(text) {
-        return text.split(/\s+/).filter(word => word.length > 0).length;
-    }
 
     createFuseIndex() {
         if (this.indexData.length > 0) {
@@ -385,108 +153,22 @@ class DocumentationIndex {
         return `Documentation related to ${topic}`;
     }
 
-    // Enable file watching for automatic index rebuilding
-    startFileWatching() {
-        if (!chokidar || isServerless) {
-            logger.info('File watching not available in serverless environment');
-            return;
-        }
 
-        if (this.watcher) {
-            logger.warn('File watcher already running');
-            return;
-        }
-
-        try {
-            this.watcher = chokidar.watch(this.contentPath, {
-                ignored: /(^|[\/\\])\../, // ignore dotfiles
-                persistent: true,
-                ignoreInitial: true, // don't trigger on initial scan
-                depth: 10 // watch subdirectories
-            });
-
-            this.watcher
-                .on('add', (path) => {
-                    logger.info(`Documentation file added: ${path}`);
-                    this.scheduleRebuild();
-                })
-                .on('change', (path) => {
-                    logger.info(`Documentation file changed: ${path}`);
-                    this.scheduleRebuild();
-                })
-                .on('unlink', (path) => {
-                    logger.info(`Documentation file removed: ${path}`);
-                    this.scheduleRebuild();
-                })
-                .on('error', (error) => {
-                    logger.error('File watcher error:', error);
-                });
-
-            logger.info(`Started file watching for documentation changes: ${this.contentPath}`);
-        } catch (error) {
-            logger.error('Failed to start file watcher:', error);
-        }
-    }
-
-    // Stop file watching
-    stopFileWatching() {
-        if (this.watcher) {
-            this.watcher.close();
-            this.watcher = null;
-            logger.info('Stopped file watching');
-        }
-    }
-
-    // Debounced rebuild to avoid too frequent rebuilds
-    scheduleRebuild() {
-        if (this.rebuildTimeout) {
-            clearTimeout(this.rebuildTimeout);
-        }
-
-        this.rebuildTimeout = setTimeout(async () => {
-            try {
-                logger.info('Rebuilding documentation index due to file changes...');
-                await this.buildIndex();
-                this.createFuseIndex();
-                logger.info(`Documentation index rebuilt with ${this.indexData.length} documents`);
-            } catch (error) {
-                logger.error('Failed to rebuild documentation index:', error);
-            }
-        }, 2000); // Wait 2 seconds after last change
-    }
-
-    // Manual rebuild method for API endpoint
     async rebuildIndex() {
-        const startTime = Date.now();
         try {
-            logger.info('🔄 Manual documentation index rebuild triggered');
-            logger.info(`   Start time: ${new Date().toISOString()}`);
-            logger.info(`   Memory usage: ${JSON.stringify(process.memoryUsage())}`);
+            const loaded = await this.loadPrebuiltIndex();
+            if (!loaded) {
+                throw new Error('Pre-built search index not found');
+            }
             
-            await this.buildIndex();
-            logger.info('📝 Build index completed, creating Fuse index...');
-            
-            this.createFuseIndex();
-            
-            const duration = Date.now() - startTime;
-            logger.info(`✅ Documentation index rebuilt with ${this.indexData.length} documents in ${duration}ms`);
-            
-            const result = { 
+            return { 
                 success: true, 
                 documentCount: this.indexData.length,
-                version: 'v2.1-clean-fix',
-                contentPath: this.contentPath,
-                duration: duration,
-                topics: Array.from(this.topics.keys()),
+                version: 'build-time-index',
                 timestamp: new Date().toISOString()
             };
-            
-            logger.info(`📤 Returning result: ${JSON.stringify(result, null, 2)}`);
-            return result;
         } catch (error) {
-            const duration = Date.now() - startTime;
-            logger.error(`❌ Failed to rebuild documentation index after ${duration}ms:`, error);
-            logger.error('Stack trace:', error.stack);
+            logger.error('Failed to reload documentation index:', error);
             throw error;
         }
     }
