+++ api/utils/logger.js
@@ -67,16 +67,15 @@ if (!isServerless) {
 
 // Create the logger
 const logger = winston.createLogger({
-    level: process.env.LOG_LEVEL || (isServerless ? 'debug' : 'info'), // Use debug level in serverless
+    level: process.env.LOG_LEVEL || 'info',
     levels,
     format,
     transports,
     exitOnError: false
 });
 
-// Keep console logging in serverless environments (Netlify Functions need console output)
-// Only remove console logging in non-serverless production environments
-if (process.env.NODE_ENV === 'production' && !isServerless) {
+// If we're in production, don't log to console
+if (process.env.NODE_ENV === 'production') {
     logger.remove(winston.transports.Console);
 }
